move interfaces  in 'type' directory
