+++ src/components/features/header.tsx
@@ -1,21 +1,24 @@
-import { ReactElement, MouseEvent, CSSProperties } from 'react';
-import { Link } from 'react-router-dom';
-import { AppRoute } from '../../const';
+import {
+  ReactElement, MouseEvent, CSSProperties, useState
+} from 'react';
 import useLogin from '../../hooks/selectors/use-login';
 import { api } from '../../services/rtk-api';
-import { getRoute } from '../../utils/common';
+import HeaderView from './header-view';
+import { messageDisplayTime } from '../../const';
+import ErrorMessage from './error-message';
 
 function Header(): ReactElement {
   const { data, isAuth, refetch } = useLogin();
   const [deleteLogout] = api.endpoints.deleteLogout.useMutation();
+  const [isShowError, setIsShowError] = useState<boolean>(false);
 
   const onSignOut = (evt: MouseEvent<HTMLElement>) => {
     evt.preventDefault();
     const apiResult = deleteLogout();
     apiResult.unwrap()
-      .catch(({ response }) => {
-      // TODO сообщение об ошибке доделать
-        throw new Error(response.error);
+      .catch(() => {
+        setIsShowError(true);
+        setTimeout(() => setIsShowError(false), messageDisplayTime);
       }).finally(() => {
         sessionStorage.removeItem('token');
         refetch();
@@ -27,47 +30,17 @@ function Header(): ReactElement {
   } as CSSProperties;
 
   return (
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link className="header__logo-link header__logo-link--active" to={getRoute(AppRoute.DEFAULT_CITY)}>
-              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width={81} height={41} />
-            </Link>
-          </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              {(isAuth
-              && (
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={getRoute(AppRoute.FAVORITES)}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper" style={avatarStyles} />
-                    <span className="header__user-name user__name">{data?.email}</span>
-                  </Link>
-                </li>
-              )
-              )}
-              <li className="header__nav-item">
-                {(isAuth
-                && (
-                  <a className="header__nav-link" href="/" onClick={onSignOut}>
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                )
-                )}
-                {(!isAuth
-                && (
-                  <Link className="header__nav-link" to={getRoute(AppRoute.LOGIN)}>
-                    <span className="header__signout">Sign in</span>
-                  </Link>
-                )
-                )}
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </header>
+    <>
+      {(isShowError
+        && <ErrorMessage text="Есть проблемы с интернет-соединением. Но, несмотря на это, мы очистили ваши данные :-)" />
+      )}
+      <HeaderView
+        isAuth={isAuth}
+        avatarStyles={avatarStyles}
+        email={data?.email}
+        onSignOut={onSignOut}
+      />
+    </>
   );
 }
