+++ src/components/features/offers-list.tsx
@@ -1,35 +1,13 @@
 import { PropsWithChildren, ReactElement, useState } from 'react';
-import { City, Hotel } from '../../data-type';
+import { Hotel } from '../../data-type';
 import CardCities from './card/card-cities';
 import Map from './map';
 import SortingDropdown from './sorting-dropdown';
-
-// TODO вынести интерфейсы в отдельный файл
-interface OffersListParams {
-  activeCity: City,
-  hotels: Hotel[],
-  sortType?: string,
-}
+import { OffersListParams } from './types';
 
 function OffersList(params: PropsWithChildren<OffersListParams>): ReactElement {
   const { activeCity, hotels } = params;
   const [activeCard, setActiveCard] = useState<Hotel>();
-
-  if (hotels.length === 0) {
-    return (
-      <div className="cities">
-        <div className="cities__places-container cities__places-container--empty container">
-          <section className="cities__no-places">
-            <div className="cities__status-wrapper tabs__content">
-              <b className="cities__status">No places to stay available</b>
-              <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
-            </div>
-          </section>
-          <div className="cities__right-section" />
-        </div>
-      </div>
-    );
-  }
 
   return (
     <div className="cities">
