+++ src/components/features/sorting-dropdown.tsx
@@ -1,17 +1,16 @@
 import { ReactElement, useState } from 'react';
-import { useHistory, useParams } from 'react-router-dom';
-import { AppRoute, SortingType } from '../../const';
-import { getRoute } from '../../utils/common';
+import { useParams } from 'react-router-dom';
+import { SortingType } from '../../const';
+import SortingDropdownBtn from './sorting-dropdown-btn';
 
 function SortingDropdown(): ReactElement {
-  const history = useHistory();
   const { city, sorting = SortingType.POPULAR.path } = useParams<{
-    city:string | undefined,
+    city: string,
     sorting: string | undefined
   }>();
   const [isOpen, setIsOpen] = useState<boolean>(false);
   const sortingArr = Object.values(SortingType);
-  const activeSorting = sortingArr.find((sort) => sort.path === sorting) ?? SortingType.POPULAR;
+  const activeSortType = sortingArr.find((sort) => sort.path === sorting) ?? SortingType.POPULAR;
 
   return (
     <form className="places__sorting" action="#" method="get">
@@ -29,7 +28,7 @@ function SortingDropdown(): ReactElement {
           }
         }}
       >
-        {activeSorting.name}
+        {activeSortType.name}
         <svg className="places__sorting-arrow" width={7} height={4}>
           <use xlinkHref="#icon-arrow-select" />
         </svg>
@@ -39,20 +38,12 @@ function SortingDropdown(): ReactElement {
         onMouseLeave={() => setIsOpen(false)}
       >
         {sortingArr.map((sortTypeName) => (
-          <li key={sortTypeName.path} className={`places__option ${activeSorting.path === sortTypeName.path ? 'places__option--active' : ''}`}>
-            <div
-              role="button"
-              tabIndex={0}
-              onClick={() => history.push(`${getRoute(AppRoute.MAIN)}${city}/${sortTypeName.path}`)}
-              onKeyDown={(ev) => {
-                if (ev.key === 'Enter') {
-                  history.push(`${getRoute(AppRoute.MAIN)}${city}/${sortTypeName.path}`);
-                }
-              }}
-            >
-              {sortTypeName.name}
-            </div>
-          </li>
+          <SortingDropdownBtn
+            key={sortTypeName.path}
+            sortType={sortTypeName}
+            activeSortType={activeSortType}
+            city={city}
+          />
         ))}
       </ul>
     </form>
