+++ src/const.ts
@@ -10,7 +10,7 @@ export enum AppRoute {
   DEFAULT_CITY = 'Paris'
 }
 
-export enum APIRoute {
+export enum ApiRoute {
   HOTELS = '/hotels',
   FAVORITE = '/favorite',
   COMMENTS = '/comments',
@@ -34,57 +34,6 @@ export enum AuthorizationStatus {
   NO_AUTH = 'NO_AUTH',
   UNKNOWN = 'UNKNOWN',
 }
-
-export const сities = [
-  {
-    location: {
-      latitude: 48.85,
-      longitude: 2.35,
-      zoom: 10,
-    },
-    name: 'Paris',
-  },
-  {
-    location: {
-      latitude: 50.9375,
-      longitude: 6.9603,
-      zoom: 10,
-    },
-    name: 'Cologne',
-  },
-  {
-    location: {
-      latitude: 50.8503,
-      longitude: 4.3517,
-      zoom: 10,
-    },
-    name: 'Brussels',
-  },
-  {
-    location: {
-      latitude: 52.370216,
-      longitude: 4.895168,
-      zoom: 10,
-    },
-    name: 'Amsterdam',
-  },
-  {
-    location: {
-      latitude: 53.5511,
-      longitude: 9.9937,
-      zoom: 10,
-    },
-    name: 'Hamburg',
-  },
-  {
-    location: {
-      latitude: 51.2277,
-      longitude: 6.7735,
-      zoom: 10,
-    },
-    name: 'Dusseldorf',
-  },
-];
 
 export const RegularExpression = {
   EMAIL: /^[^\s@]{3,}@[^\s@]{2,}\.[^\s@]{2,}$/,
